requirements.txt
flask
google-genai
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
python-telegram-bot==21.*
apscheduler
pytz
